feat: create component specific for the observer
