+++ src/frontend/src/routes/(app)/+layout.svelte
@@ -9,6 +9,7 @@
 	import { authSignedIn } from '$lib/derived/auth.derived';
 	import Loaders from '$lib/components/core/Loaders.svelte';
 	import NoLoaders from '$lib/components/core/NoLoaders.svelte';
+	import CollapseObserver from '$lib/components/hero/CollapseObserver.svelte';
 
 	let route: 'transactions' | 'tokens' | 'settings' = 'tokens';
 	$: route = isRouteSettings($page)
@@ -21,13 +22,18 @@
 
 	let cmpLoaders: ComponentType;
 	$: cmpLoaders = $authSignedIn ? Loaders : NoLoaders;
+
+	let collapse = false;
 </script>
+
+<CollapseObserver bind:collapse />
 
 <Hero
 	usdTotal={route === 'tokens'}
 	summary={route === 'transactions'}
 	more={route === 'transactions'}
 	actions={route !== 'settings'}
+	{collapse}
 />
 
 <main class="pt-12">
